move table_reader code to table_reader.go
